+++ my-app/app/my-bugs/page.tsx
@@ -1,0 +1,5 @@
+import { MyBugsView } from "@/features/my-bugs/views/MyBugsView";
+
+export default function MyBugsPage() {
+  return <MyBugsView />;
+}